Put the StreamConverters into a separate spec
